requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
python-dateutil
matplotlib
streamlit
requests
python-dotenv